requirements.txt
fastapi==0.115.12
motor==3.7.1
pydantic==2.11.4
pymongo==4.13.0
setuptools==80.1.0
starlette==0.46.2
uvicorn==0.34.2
python-dotenv==1.1.0
python-jose==3.4.0
pytz==2025.2
redis==6.1.0
bcrypt==4.3.0
python-multipart==0.0.20
requests==2.32.3
python-telegram-bot==22.3
APScheduler==3.11.0
pyotp==2.9.0
numpy==1.26.4
httpx==0.27.2
aiohttp==3.10.5
